refactor: update import paths to use main.js and index.js for consistency
