+++ src/index.ts
@@ -1,2 +1,2 @@
-export { default as JasonDB } from "./core/core.js";
-export type { BaseDocument } from "./types/type.js";
+export { default as JasonDB } from "./core/main.js";
+export type { BaseDocument } from "./types/index.js";
