+++ test/collections.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { rm } from "node:fs/promises";
 import path from "node:path";
-import JasonDB from "../src/core";
-import type { BaseDocument } from "../src/type";
+import JasonDB from "../src/core/main";
+import type { BaseDocument } from "../src/types";
 
 interface TestUser extends BaseDocument {
 	id: string;
