+++ test/create.test.ts
@@ -1,7 +1,7 @@
 import { rm } from "node:fs/promises";
 import path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import JasonDB from "../src/core";
+import JasonDB from "../src/core/main";
 import type { TestCollections, TestPost } from "./types";
 
 const testFilename = "test_create_db";
