+++ test/update.test.ts
@@ -1,8 +1,8 @@
-import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { rm } from "node:fs/promises";
 import path from "node:path";
-import JasonDB from "../src/core";
-import type { BaseDocument } from "../src/type";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
+import JasonDB from "../src/core/main";
+import type { BaseDocument } from "../src/types";
 
 interface TestUser extends BaseDocument {
     id: string;
@@ -44,7 +44,7 @@ describe("USER tests", () => {
         const users = db.collection("users");
         await expect(
             users.update("non-existent-id", { age: 40 }),
-        ).rejects.toThrowError("Failed to acquire lock"); 
+        ).rejects.toThrowError("Failed to acquire lock");
     });
 });
